attached_assets/Pasted-Error-Element-type-is-invalid-expected-a-string-for-built-in-components-or-a-class-function-for-1748291594186.txt

Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.
    at renderElement (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6043:9)
    at renderNodeDestructiveImpl (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6104:11)
    at renderNodeDestructive (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6076:14)
    at renderNode (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6259:12)
    at renderChildrenArray (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6211:7)
    at renderNodeDestructiveImpl (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6141:7)
    at renderNodeDestructive (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6076:14)
    at renderNode (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6259:12)
    at renderHostElement (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:5642:3)
    at renderElement (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:5952:5)
    at renderNodeDestructiveImpl (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6104:11)
    at renderNodeDestructive (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6076:14)
    at renderIndeterminateComponent (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:5785:7)
    at renderElement (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:5946:7)
    at renderNodeDestructiveImpl (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js:6104:11)
